fix(proxy): split Set-Cookie on newline in writeResponseFromBuffer
Playwright's response.allHeaders() joins multiple Set-Cookie values
with \n into a single map entry. writeResponseFromBuffer was writing
the raw string directly, embedding literal newlines in the HTTP
response. HTTP/1.1 clients that validate header syntax (Go's net/http,
Java's HttpClient) reject lines like

  session-id-time=123l; Expires=Tue, 01 Jan 2030 ...; Path=/

as malformed because they have no header-name: prefix.

Fix: split on \n and emit each cookie as its own Set-Cookie: line,
matching RFC 6265 which explicitly forbids comma-folding for cookies.

Adds two unit tests using a PassThrough stream as a socket mock.
